refactor(kikan,api): extract ProfileDb + activity_writer prep [#507]
V6c infra prep (a1 + a2) — pure platform refactor, no behavior change.

- Move `ProfileDb` type + `FromRequestParts` impl from services/api into
  `crates/kikan/src/profile_db.rs`, re-exported as `kikan::ProfileDb`.
  Rejection body emits the platform-wide ErrorBody wire shape
  (`{"code":"internal_error","message":...,"details":null}`) inline,
  avoiding a kikan-types dep (which already depends on kikan for
  `SetupMode` — the edge can't reverse).
- Keep `profile_db_middleware` in services/api/src/profile_db.rs — it's
  welded to `SharedState` (reads `state.db_for(mode)` and
  `state.active_profile`) and can't move into platform code until
  `AppState` lands behind the Graft boundary.
- Update imports in services/api/src/{activity,auth,auth/reset,auth/recover,customer}
  to pull `ProfileDb` from `kikan`.
- Add `activity_writer: Arc<dyn kikan::ActivityWriter>` to AppState,
  initialized with `Arc::new(kikan::SqliteActivityWriter::new())`. This
  prepares V6c.b2 to wire the customer handler's router state without a
  second AppState plumbing pass.

Why: router state is captured at `with_state(...)` time and shared
across requests — a DB handle there would break profile switching
silently. Per-request DB via `ProfileDb` extractor preserves the
seamless demo ↔ production promise. Future Graft migration (S2.5+)
renames `activity_writer` from AppState to EngineContext without
behavior change.

Gates: cargo check --workspace green; cargo clippy -p kikan
--all-targets -D warnings green; kikan profile_db tests (3) pass;
services/api lib tests (126) pass; I1–I5 green; R13 green.
