+++ crates/kikan/src/lib.rs
@@ -9,6 +9,7 @@ pub mod error;
 pub mod graft;
 pub mod middleware;
 pub mod migrations;
+pub mod profile_db;
 pub mod tenancy;
 
 pub use activity::{ActivityLogEntry, ActivityWriter, SqliteActivityWriter};
@@ -20,4 +21,5 @@ pub use error::{
 };
 pub use graft::{Graft, SelfGraft, SubGraft};
 pub use migrations::{GraftId, Migration, MigrationRef, MigrationTarget};
+pub use profile_db::ProfileDb;
 pub use tenancy::{ProfileId, SetupMode, Tenancy};
