+++ services/api/src/customer/mod.rs
@@ -16,7 +16,7 @@ use crate::SharedState;
 use crate::auth::AuthSessionType;
 use crate::error::AppError;
 use crate::pagination::PaginationParams;
-use crate::profile_db::ProfileDb;
+use kikan::ProfileDb;
 
 pub fn router() -> Router<SharedState> {
     Router::new()
